refactor: Clarify app structure
