+++ app/api/services/image_sinks/s3_sink.py
@@ -19,7 +19,7 @@ from typing import TYPE_CHECKING, Any, Protocol
 from pydantic import AnyUrl
 
 from app.api.services.image_sinks.base import ImageSinkError, StoredImage
-from app.utils.logging import build_log_extra
+from app.observability.logging import build_log_extra
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
