+++ app/api/services/pairing_client.py
