+++ app/observability/__init__.py
@@ -1,0 +1,1 @@
+"""Logging, tracing, and telemetry wiring."""
